Add files via upload
This file is the file a modified the sample.py to fit our cs418 data extraction work. Please replace the row 48, API_KEY as your own api key. I also add some empty list in the main function, you can see starting from row 178. The containers are to contain the information we need. each time, we can only has 50 record for the data. it is the limit of the yelp api. Try to base on that, to solve the problem.